Code refactoring
Move any CPM related functions from main.py to cpm.py. The idea is to
have cpm.py act as a library, and main.py act as a command-line
interface.
This makes it easy to create additional interfaces, such a web
interface, without implementing some things all over again.
The goal is to have a standalone cpm.py module that can be used by
anyone that wants to add the CPM algorithm in a project.