Add optimum solution feature and fix images bug
Modify the cpm.get_results method to find and return the optimum
solution of the project. Optimum solution is defined as the combination
of the minimum total cost in a minimum project duration. Modify both
interfaces (cli and web) to make use of this feature.
Fix a bug related with the loading of the images in the results of the
web interface. Because every set of generated images shares the same
filename pattern (i.e., network-0.png, network-1.png, etc.), the cache
mechanism of the browser was not bothering to reload the images. This
led to a situation where the user was getting some images from the
previous project that she run, and some from the current one.